requirements.txt
fastapi==0.117.1
pydantic==2.11.9
requests==2.32.5
uvicorn==0.36.0
python-dotenv==1.1.1
Pillow==10.4.0
supabase
google-genai
gtts>=2.5.0
openai
reportlab>=4.0.0
slowapi>=0.1.9
pyjwt>=2.8.0
cryptography>=41.0.0
bleach>=6.1.0
stripe>=7.0.0

# Optional: For enhanced virus scanning (requires ClamAV daemon)
# clamd>=1.0.2